+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to the "flutter-intl" extension will be documented in this f
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.2.0 - 2020-03-16
+
+- Add support for locales with script code
+
+- Fix locale loading issue when country code is not provided
+
 ## 1.1.0 - 2020-02-04
 
 - Make main locale configurable